refactor(container): lift the exec contract out of client.ts
`client.ts` had crossed 1,000 lines carrying six responsibilities. The
exec block is a clean seam: it is a self-contained wire protocol whose
only tie to the rest of the client is a handle's `fetch`, which both
factories already pass in. Declaring that dependency structurally
(`ExecFetch`) rather than as `ContainerHandle["fetch"]` keeps the import
one-way.

The bounded reader moves to its own module because both sides need it and
neither owns it — the cold-start retry scans a response prefix for a
provisioning sentinel, the exec contract reads a whole capped result.

Pure move: no behaviour change, no public surface change.
